restructured based on grouping by features/route
