+++ src/components/Rapid/Display.js
