+++ src/components/Rapid/Display.scss
