+++ src/components/Rapid/Eddy.js
