+++ src/components/Rapid/NextRapid.scss
