+++ src/components/interface/Global.js
@@ -8,9 +8,9 @@ import {
   useParams,
   useRouteMatch,
 } from "react-router-dom";
-import SearchBar from "../SearchBar";
-import RiverList from "../../../river-data/RiverList";
-import { Exists, UnderConst, DoesNotExist } from "../RiverCards/RiverCards";
+import SearchBar from "./SearchBar";
+import RiverList from "../../river-data/RiverList";
+import { Exists, UnderConst, DoesNotExist } from "./RiverCards";
 import GithubCorner from "react-github-corner";
 
 class Global extends React.Component {
