+++ src/components/river/GenericToggle.js
