+++ src/components/river/MapLabel.js
@@ -9,7 +9,7 @@ import {
   useParams,
   useRouteMatch,
 } from "react-router-dom";
-import idParser from "../tools/idParser";
+import idParser from "../../tools/idParser";
 
 const MapLabel = (props) => {
   const pointerDirection = (pointerDirection) => {
