+++ src/components/river/Slider.js
